# Performance and memory optimizations