@fatkodima 2023-03-06
> Is the main point avoiding `String` object allocations?

The main point was to reduce the allocated memory, yes. Allocated strings and match data objects were reduced. And additionally a few methods were optimized as a side effect.